project/requirements.txt
Django==2.2.17
django-tailwind==0.9.0
djangorestframework==3.12.2
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1
